Fix reservation

+++ src/main/java/ma/karflow/feature/billing/dto/InvoiceResponse.java
@@ -2,6 +2,7 @@ package ma.karflow.feature.billing.dto;
 
 import ma.karflow.feature.billing.enums.InvoiceStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -10,13 +11,13 @@ import java.util.UUID;
 public record InvoiceResponse(
         UUID id,
         String invoiceNumber,
-        double subtotal,
-        double taxRate,
-        double taxAmount,
-        double discount,
-        double totalAmount,
-        double totalPaid,
-        double remainingAmount,
+        BigDecimal subtotal,
+        BigDecimal taxRate,
+        BigDecimal taxAmount,
+        BigDecimal discount,
+        BigDecimal totalAmount,
+        BigDecimal totalPaid,
+        BigDecimal remainingAmount,
         InvoiceStatus status,
         LocalDate dueDate,
         LocalDate paidDate,
