+++ src/main/java/ma/karflow/feature/rental/dto/RentalResponse.java
@@ -2,6 +2,7 @@ package ma.karflow.feature.rental.dto;
 
 import ma.karflow.feature.rental.enums.RentalStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -11,10 +12,10 @@ public record RentalResponse(
         LocalDate startDate,
         LocalDate endDate,
         LocalDate actualReturnDate,
-        Double mileageBefore,
-        Double mileageAfter,
-        double deposit,
-        double totalAmount,
+        BigDecimal mileageBefore,
+        BigDecimal mileageAfter,
+        BigDecimal deposit,
+        BigDecimal totalAmount,
         RentalStatus status,
         String notes,
         UUID vehicleId,
