+++ src/main/java/ma/karflow/feature/vehicle/dto/CategoryResponse.java
@@ -1,11 +1,12 @@
 package ma.karflow.feature.vehicle.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CategoryResponse(
         UUID id,
         String name,
         String description,
-        double dailyRateMultiplier
+        BigDecimal dailyRateMultiplier
 ) {
 }
