+++ src/main/java/ma/karflow/feature/vehicle/dto/VehicleResponse.java
@@ -2,6 +2,7 @@ package ma.karflow.feature.vehicle.dto;
 
 import ma.karflow.feature.vehicle.enums.VehicleStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,8 +10,8 @@ public record VehicleResponse(
         UUID id,
         String licensePlate,
         String color,
-        double mileage,
-        double dailyRate,
+        BigDecimal mileage,
+        BigDecimal dailyRate,
         String photoUrl,
         VehicleStatus status,
         VehicleModelResponse vehicleModel,
